Signup -> then, make a empty accounts model -> stored outbox at first
Using Kafka, send data from outbox to accounts domains (so, user can make an account information successfully)

+++ accounts/domain/value_objects.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 @dataclass(frozen=True)
 class Country:
@@ -7,7 +8,7 @@ class Country:
 
 @dataclass(frozen=True)
 class AnnualIncome:
-    amount: int
+    amount: Decimal
 
     def __post_init__(self):
         if self.amount <= 0:
